+++ internal/domain/config/shell/dto.go
@@ -1,0 +1,7 @@
+package shell
+
+type Shell struct {
+	After   string `yaml:"after"`
+	Before  string `yaml:"before"`
+	Enabled bool   `yaml:"enabled" default:"true"`
+}