+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented here.
 ## [1.9.0] - 2025-11-16
 - Add support docker
 - Fix issue create dump file in only main dir
+- Add support shell script before and after backup
 
 ## [1.8.3] - 2025-11-15
 - Fix bug with create dump mongodb